[ForLoop] Port ForLoop TMP to C++11
The ForLoop TMP used to put a number of restrictions on the arguments to
apply(), mainly due to the limited capabilities of C++01 wrt. to
argument forwarding.

But now we have C++11, so we can throw away all that awkward code and
replace it with a single forwarding function that supports arbitrary
numbers of arbitrary combinations of lvalues and rvalues - all hail
variadic templates and perfect forwarding!