[parametertree] keep track of the current prefix
in order to improve error messages further we introduce a new member
variable prefix_. Sometimes it is important to pass subtrees around.
In this case it is difficult to interpret error messages, as the user
doesn't know in which subtree the problem occured. Now a Parametertree
knows its prefix and can tell it to the user. This way the report
method is simplified and the throws are now more readable.