[dunecontrol] fix make command when using cmake (FS#1572)
We fix the option handling when calling "cmake --build", by prepending '--' to the command
line parameters, in order to separate cmake and build parameters, iff the user didn't alrready do
it himself so. Note that this means we assume that the users generally want to pass
parameters to the build command and not to cmake itself.