Do not convert hyphens to underscores in the name of the example .cc program
Each module constructed by duneproject contains a little Hello-world example
program.  The name of the program used to be the name of the module with .cc
appended AND all hyphens replaced by underscores.  Hence a module dune-foo
would containt the program dune_foo.cc.  This patch removes that displacement.
Rationale:
- I see no reason for the replacement in the first place
- File names with hyphens appear to be much more common in Dune world than
  file names with underscores.