[m4] fix mixup when using autools/cmake 2.4 + an installed 2.3
Ich a module is not updated to cmake yet, we can have the situation, that the core modules
are built using cmake, but used by autotools. When trying to detect a dune library, we either
explicitly stated the /blablabla/libdunefoo.la file if the the dependency was built using autools,
or used the '-ldunefoo -L/blablabla/' options. This failed with very subtle errors if dune was
also installed globally, e.g. a 2.3 from your favorite distribution. We fix this now in the same
way as we fixed it for autotools. If the module is local, we explicitly state the lib...
either as .la or as .a.