[bugfix] Only communicate on non-empty interfaces in VariableSizeCommunicator
Before this commit communcation would be initiated even on empty
interfaces. While this does not lead to overwriting data or producing
segmentation faults, it does lead to accessing the first component of
empty vectors to get a pointer to pass to MPI functions. This produces
error output with valgrind and other memory debuggers. Therefore with this
commit we will first check whether the interface contains values and do
nothing if it doesn't. We also augmented the test to catch the case wher only
some processes have an empty interface.