[Release][CMake] Add dune_target_enable_all_packages()
This patch adds a new function dune_target_enable_all_packages() that
works like dune_enable_all_packages(), but only acts on the given list
of targets.

This function is mostly needed to be able to use the new
dune_enable_all_packages() feature in a DUNE module that defines its own
libraries. Specifically, the "nice" default way of handling those
libraries only works on CMake 3.1+. To be compatible with older CMake
versions, the library has to be manually added in the top-level
CMakeLists.txt file before the call to dune_enable_all_packages(). But
the library also needs to get all the package flags from the DUNE
modules the current one depends on, which can be done using the new
function added by this patch.