[Release][CMake][Bugfix] Fail hard if we find nullptr, but not std::nullptr_t
We do not really support this combination of C++11 features, and the
only time it should occur is when the user has a recent Intel Compiler
and an outdated GCC (which the Intel Compiler uses for the C++ standard
library). In that case, we emit a helpful error message at configuration
time.

(cherry picked from commit c3e637abffbe60c12aa57df84fdef797b1cbc5be)
Signed-off-by: Steffen Müthing <muething@dune-project.org>