[CMake] Replace outdated external TBB test
This patch replaces the external FindTBB module with a new
implementation that actually works correctly with recent versions of TBB
and those bundled with Intel compilers. It also solves the license
problem of the old test, which was MIT.

The new tests use the COMPONENTS feature to enable or require the TBB
preview build and / or the TBB allocator add-on library. Usage of the
debug version is no longer coupled to the CMake build type (which
doesn't make sense in this setting), but to a new option TBB_DEBUG.

The input and output variables have changed as well, take a look at the
file for further details.

The patch also merges FindTBB and DuneTBB and adds an additional final
compile test to make sure the detected configuration actually works.