[bugfix] make fvectortest compile and make ScalarOrderingTest pass for GMPField
the first problem was that I accidentially enabled the ScalarOrderingTest, but this forced me
to investigate further and I found out how this works, or not works.

We have to add the comparison operators to FieldVector, as the cast operator does not always
match, e.g. in the case of GMPField.
(cherry picked from commit e0ac8d9a263772437c518d55f90da93204d0a4b7)

Signed-off-by: Carsten Gräser <graeser@dune-project.org>