[CMake] Require CMake 2.8.12 for dune_enable_all_packages()
dune_enable_all_packages has to update the compiler options, but the
command compile_options only exists in CMake 2.8.12+. So we add a hard
requirement for that CMake version to the function.

I only ever intended this feature to work for CMake >= 2.8.12, so this
doesn't really change anything, except for giving the users a more
useful diagnostic.