[release] MPITraits: Use std::uint16_t for bigunsignedint<k>
The storage of bigunsignedint was changed to std::uint16_t in commit
541b3dfceb6f53f34d8f6ded5a27289b13e9afbb. This implements the same
change for the MPITraits<bigunsignedint<k>>

Note that this assumes that
  MPITraits<std::uint16_t>
is defined. This is the case as long as uint16_t maps to one of the
standard integer types (unsigned char/short/int/long).

Thanks to Markus for pointing this out.

Reference: https://dune-project.org/flyspray/index.php?do=details&task_id=1657#comment6393