[bugfix][Release] Initialize index with an unsigned value
With the recent changes to bigunsignedint, the index cannot
be initialized with a negative value anymore. We change this
to 1 to overcome the problem.

This commit together with the recent bugfixes by Ansgar
fixes the AMG code on master.

(cherry picked from commit bf0239472d01319211f5c0f7b37bf4938e9e29a2)
Signed-off-by: Carsten Gräser <graeser@dune-project.org>