[IteratorRange][Release] Make IteratorRange default-constructible and copyable
IteratorRanges are convenient for storing ranges of iterators into a
larger container, e.g. for the per-codim GeometryTypes in an
IndexSet. Unfortunately, you can't just create an array of
IteratorRanges right now because they can neither be copied, nor are
they default-constructible.

This patch fixes that problem and also adds a second typedef
const_iterator to make IteratorRange work in contexts where the
iterating code isn't using range-based for and assumes a const object.