[bugfix,cmake] Correctly remove dune-common_MODULE_PATH from CMAKE_MODULE_PATH
This has to be done as each module puts dune-common_MODULE_PATH into
CMAKE_MODULE_PATH to find the DUNE's cmake macros. This should be removed
when building the complete CMAKE_MODULE_PATH module path (in reverse order
of the dependency). Unfortunately, the conditional to check whether
the variable dune-common_MODULE_PATH is set was incorrect. This commit fixes
this.

I hope that this fixes FS 1686
https://dune-project.org/flyspray/index.php?do=details&task_id=1686

Please merge to the release