[release] Make finding *.la files non-fatal.
Actually, having these files lying around does not hurt us.
In addition there are downstream modules that especially create
these for compatibility reasons with autotools. For these
reconfiguring a module using dunecontrol currently aborts with an error

Therefore this commit resorts to just issuing a warning and proceeding
with the compilation instead of aborting with an error.