[cmake] add SuiteSparse support, replace UMFPACK version
This patch adds the cmake findpackage script which locates Tim Davis' SuiteSparsae collection.

If a certain setting of the SuiteSparse library should be used it can be found using
find_package( SuiteSparse COMPONENETS <list-of-components> )
with a specific list of components.
Existence of components can be check using the SUITESPARSE_<comp>_FOUND/HAVE_SUITESPARSE_<comp> cmake variable.

The UMFPack search routines are redirected to the more general FindSuiteSparse script.