[bugfix] make AddRefTypeEvaluator work in tranformTuple
As documented, tranformTuple should support AddRefTypeEvaluator. As
std::make_tuple will remove references (using std::decay) from the
funtor's return type, the returned tuple would contain copies of the
values. This is not the intended behavior.

This patch adds AddRefTypeEvaluator to the utilitytest and fixes the
return type of transformTuple.