[CMake] Introduce DUNE_BUILD_TESTS_ON_MAKE_ALL
As talked about on the developer meeting, a configure
switch should allow to build the tests during make all.
You can either set it through your opts file or activate
it on a per module level. If not active, you have to run

make build_tests && make test

to run the test suite.