[CMake] Implement a COMPILE_ONLY option for dune_add_test
Set if the given test should only be compiled during `make build_tests`,
but not run during `make test`. This is useful if you compile the same
executable twice, but with different compile flags, where you want to assue that
it compiles with both sets of flags, but you already know they will produce the
same result.