Merge branch 'feature/gitlab-readable-contribution-and-licensing-guide' into 'master'
GitLab readable contribution guide, license and README

Adding the `.md` extension enables rendering of the files, which is much nicer.

Also ported the contribution guide from the User Wiki.

Renamed COPYING to LICENSE.md because that is the name that GitLab looks for to display it nicely on the project starting page.

If this one is accepted, I will do the same for the other core modules.

See merge request !5
(cherry picked from commit 633391ff25d9ddc7294e773e876949d687ae78ea)

Signed-off-by: Steffen Müthing <muething@dune-project.org>