Do not guard DUNE_UNUSED_PARAMETER
If the right defines are set, the variables in question will actually be
used. This does not seem to be a problem. The gcc documentation has this
to say about the "unused" attribute:

This attribute, attached to a variable, means that the variable is meant
to be *possibly* unused. GCC will not produce a warning for this
variable. (emphasis added)