[cmake] extract bottom block from config.h.cmake
For dune-grid a bottom section is needed to place the Grid Type magic at the bottom of the config.h
file in derived modules.

In this patch a /* begin bottom */ ... /* end bottom */ block from each module's config.h.cmake file and placed
at the end of the collected config.h file.

This resolves the problems from FS#1713.