[bugfix] GCC was giving too many deprecation warnings for enable_if
I still don't understand why, but GCC was spilling deprecation warnings
for Dune::enable_if even if we were only using std::enable_if.

Thus I removed the deprecation of the enable_if struct and only
deprecate the type within. That should still be enough as this
deprecation warning gets triggered when enable_if is used to aid in
overload resolution and the boolean parameter of enable_if is true.