[Test] Disable debugallocator tests on OS X
Those tests have been failing for as long as I can remember, I think
they fail due to some Linux <-> BSD differences, and the feature is too
obscure to warrant extensive fixing.

The patch also adds a note to the DebugAllocator documentation that
warns of the OS X problem and suggests alternatives.