Make testdebugallocator_fail{1,4,5} pass on OS X
mprotect violations are required to lead to a SIGSEGV on Linux.
OS X is more permissive, allowing for both SIGSEGV and SIGBUS.

And practice, the tests generate a SIGBUS for me, causing the tests to
fail. Confusingly, ctest describes the problem as

  Exception: SegFault

which I assume they use as a catch-all term.