[CMake] Don't pollute build-specific CXX_FLAGS with global compiler flags
The C++ version flag check used to set the -std=... flag for both the
basic CMAKE_CXX_FLAGS and the build-specific variants of it. That is
unnecessary because CMAKE concatenates the base variable and the
build-specific variant (if a build type has been set by the
user). Having both is actually rather unhelpful, as any downstream
changes will also have to be done to the complete set of variables.