[CMake] Add function to allow modules to require a minimum C++ standard
This patch adds a new function dune_require_cxx_standard() that any Dune
module can call to require support for at least a given C++
standard. If the compiler doesn't meet the requirements, the function
will report the required and the actually available standard versions
and abort with a fatal error.

Moreover, it knows about CXX_MAX_STANDARD and will tell the user if the
value of that variable is below the requirements. This avoids desperate
users that have a shiny beta of GCC 6 with C++17 support wondering why
their own module using shiny C++17 concepts stubbornly fails to
build... ;-)