[release] Find METIS even if it is actually the static Scotch library.
The Scotch library on Debian is rather broken. There are several libraries
missing for the version with long, int32, int64 support. If one
wants to use them, then one needs to use the static libraries and provide
the correct library with the metis bindings. Unfortunately, our METIS test
assumes that the library is self contained which breaks on Debian Jessie.

This commit fixes this. If METIS_PartGraphKway is not found with the METIS
library alone then we try to find the corresponding Scotch helper libraries
and try to find the function again.