[release] Find ParMETIS even if it is actually the static PT-Scotch library.
The PT-Scotch library on Debian is rather broken. There are several libraries
missing for the version with long, int32, int64 support. If one
wants to use them, then one needs to use the static libraries and provide
the correct library with the parmetis bindings. Unfortunately, our ParMETIS test
assumes that the library is self contained which breaks on Debian Jessie.
This commit fixes this. If parmetis_v3_partkway is not found with the ParMETIS
library alone then we try to find the corresponding PT-Scotch helper libraries
and try to find the function again.