Merge branch 'feature/fix-cmake-documentation-without-doxylocal' into 'master'
[cmake][bugfix] Make CMake Doxygen creation work for modules without Doxylocal

When creating Doxygen documentation, CMake currently assumes that all
modules have a Doxylocal file, which is in fact only present in
dune-common right now. That file is then used as both a dependency and a
file reading input, causing CMake errors.

This patch fixes the problem by testing for the presence of Doxylocal
and only including it in the build process if it was found.

See merge request !61
(cherry picked from commit 4b1fb7c8eb51697e0de593c6a2fa6675f05879ea)

Signed-off-by: Steffen Müthing <muething@dune-project.org>