[bugfix][concept] Fix tupleEntriesModel() check
The old version did not work for tuples with more than two entries
because recursive function templates don't work this way. Now we
simply use a TMP reduce on the tuple.

This also fixes the problem that composite bases in dune-functions
with more than two factors did not compile due to a failing concept
check caused by this issue.