[Std] Update implementation of integer sequences from C++14
Most of the time we can rely on std::integer_sequence et. al. being
available, as GCC 4.9+ runs in C++14 mode, so the new version tries to
import the functionality from the standard library.

If it is not available, we now use a fully compliant
implementation (i.e. index_sequence is now an alias template of
integer_sequence, and the various construction helpers are now also
alias templates instead of functions).