Add helper wrap_or_move()
This is a variant of stackobject_to_shared_ptr that
can store L- and R-value references. If you pass
an R-value reference it behaves like stackobject_to_shared_ptr.
If you pass an L-value reference it will be moved
into a new object created via make_shared.

The main application of this is the following situation:
If you have a function

  template<class T>
  auto foo(T&& t);

it can be used with temporaries and L-values. If however,
this should return an object referring to t, you can no longer
pass a temporary unless you prolong its life-time. The latter
can be done conveniently using wrap_or_move().