Support multiple DUNE libs in different directories when exporting.
Previously, issuing multiple dune_add_library calls would only
work if subsequent calls to dune_add_library were issued in the same
directory or below. If the calls happened in different directories
then only the last library would have beem exported.
With this commit we can use the APPEND parameter to prevent
overwriting previous libraries.

If all calls are in the same directory or below, then APPEND can be
omitted. But that is not recommend.