[cmake] Add a CMAKE_GUARD option to dune_add_test
A list of conditions can be passed to the argument. CMake
will evaluate these conditions, and if one of them does not
evaluate to TRUE, the test sources will be replaced by a dummy
source, that returns 77.

This option should be used instead of guarding the call to
dune_add_test with a cmake if-clause. The idea behind it is
to always give the user an idea of which tests are available,
but have not been run on his system.

The commit also applies the new argument to the only occurence in
dune-common.