[bugfix] Use const references in Hybrid::Imp::equals()
If T1 and T2 are not defined by 'const T1&', they will themselves
be of reference type. Hence T2::value does not exist and the
equals() overload for the static evaluation is not selected.
Then the dynamic equals() overload is used. This will give the
same result, but the compiler will instantiate the if branch
with the identity if the result is 'false'.

This could be fixed using std::decay_t. But since comparison
should be a const operation anyway, using a const reference
is cleaner here.