[cleanup] Avoid unused variable warning
We could cast the std::initializer_list to void to avoid the
warning. But extracting this as a function makes the code more
readable.

Maybe evaluateFoldExpression() should be moved outside of Imp::
because this little trick nicely helps to avoid a template recursion
and may help to simplify other places, too.