[test] Work around incomplete C++14 support in gcc-4.9
In C++14 constexpr does not imply const in contrast to C++11.
Since this is not implemented in gcc-4.9 you cannot have
a non-const constexpr function. Hence we need a non-constexpr
operator[] if we want a non-const one.

Notice that the const of the constexpr version is not needed
in gcc-4.9 but in newer versions implementing this correctly.