add_dune_test: Handle complex expressions in `CMAKE_GUARD`
The check `if(NOT ${condition})` does not behave as expected when
`condition` is a complex string expression such as `A OR B`. It does
however work when `condition` is a list.

Therefore convert `condition` to a list by calling `separate_arguments`.

Note that this change also adds missing parenthesis around `condition`
which also results in wrong results for complex expressions.

Closes issue #39.