[hybrid] Change return type of Hybrid::ifElse from void to auto
Since Hybrid::ifElse() is essentially an if statement, having
support for return values looks strange. However, this is not
visible if you don't use it and allows to at least partially
emulate returning from functions inside of ifElse()-branches.

Before you could not return from those branches, because they
are functions themselves that 'eat up' the return statement.
Now you can at least return values for cases where both
branches would return.

  template<class T
  auto fooOrBar(T t) {
    return ifElse(models<HasFoo, T>(), [&](auto id) {
      return id(t).foo();
    }, [&](auto id) {
      return id(t).bar();
    });
  }

Mind the additional "return" in front of ifElse() that
which is needed to forward the returned value
Notice that it's not required that both branches have the same
return type if isElse() is in static mode.