[parametertree][bugfix] Fix handling of keys that appear as both subtrees and values
This patch adds additional checks to the ParameterTree that avoid the
occurence of keys with both a subtree and a value attached, which can
theoretically happen due to the internal implementation of the
ParameterTree. As discussed in #33, this is undesirable as the resulting
trees cannot mapped to the Python dicts used in dune-python.