make dunecontrol compatible with git worktrees
In linked worktrees generated with `git worktree`, .git is a file
instead of a directory.

Since dunecontrol was only checking whether .git exists as a directory,
it didn't recognize linked worktrees as git repositories. This problem
can simply be fixed by additionally checking if .git exists as a regular
file.